+++ packages/datatrak-web/src/api/queries/useSurveyScreenComponents.ts
@@ -2,7 +2,7 @@
  * Tupaia
  *  Copyright (c) 2017 - 2023 Beyond Essential Systems Pty Ltd
  */
-import { groupBy } from 'lodash';
+import { groupBy, sortBy } from 'lodash';
 import { useQuery } from 'react-query';
 import { DatatrakWebSurveyScreenComponentsRequest as ScreenComponentsRequest } from '@tupaia/types';
 import { get } from '../api';
@@ -14,6 +14,7 @@ export const useSurveyScreenComponents = surveyCode => {
       get(`surveys/${surveyCode}/surveyScreenComponents`),
   );
 
-  const mappedData = groupBy(data, 'surveyScreenScreenNumber');
+  const mappedData = groupBy(sortBy(data, 'componentNumber'), 'surveyScreenScreenNumber');
+
   return { ...query, data: mappedData };
 };
