+++ packages/datatrak-web/src/features/Survey/SurveyContext.tsx
@@ -5,20 +5,9 @@
 
 import React, { Dispatch, createContext, useContext, useReducer } from 'react';
 import { useParams } from 'react-router-dom';
-import { sortBy } from 'lodash';
 import { SurveyParams, SurveyScreenComponent } from '../../types';
 import { useSurveyScreenComponents } from '../../api/queries';
-
-const convertNumberToLetter = (number: number) => {
-  const alphabet = 'abcdefghijklmnopqrstuvwxyz';
-  if (number > 25) {
-    // occasionally there are more than 26 questions on a screen, so we then start at aa, ab....
-    const firstLetter = alphabet[Math.floor(number / 26) - 1];
-    const secondLetter = alphabet[number % 26];
-    return `${firstLetter}${secondLetter}`;
-  }
-  return alphabet[number];
-};
+import { formatSurveyScreenQuestions } from './utils';
 
 type SurveyFormContextType = {
   formData: Record<string, any>;
@@ -29,6 +18,7 @@ type SurveyFormContextType = {
   screenHeader?: string;
   displayQuestions: SurveyScreenComponent[];
   sideMenuOpen?: boolean;
+  isReviewScreen?: boolean;
   surveyScreenComponents?: Record<number, SurveyScreenComponent[]>;
 };
 
@@ -88,13 +78,9 @@ export const SurveyContext = ({ children }) => {
   const screenNumber = parseInt(params.screenNumber!, 10);
   const { data: surveyScreenComponents } = useSurveyScreenComponents(surveyCode);
 
-  const activeScreen = sortBy(surveyScreenComponents[screenNumber!], 'componentNumber');
+  const activeScreen = surveyScreenComponents[screenNumber!] ?? [];
   const numberOfScreens = Object.keys(surveyScreenComponents).length;
   const isLast = screenNumber === numberOfScreens;
-
-  const nonInstructionQuestions = activeScreen.filter(
-    question => question.questionType !== 'Instruction',
-  );
 
   // for the purposes of displaying the screen number, we don't want to count screens that only have instructions
   const screensWithQuestions = Object.values(surveyScreenComponents).filter(screen => {
@@ -111,19 +97,18 @@ export const SurveyContext = ({ children }) => {
     : -1;
   // If the first question is an instruction, don't render it since we always just
   // show the text of first questions as the heading. Format the questions with a question number to display
-  const displayQuestions = (activeScreen.length && activeScreen[0].questionType === 'Instruction'
-    ? activeScreen.slice(1)
-    : activeScreen
-  ).map(question => {
-    const questionNumber = nonInstructionQuestions.findIndex(
-      nonInstructionQuestion => question.questionId === nonInstructionQuestion.questionId,
-    );
-    if (questionNumber === -1) return question;
-    return {
-      ...question,
-      questionNumber: `${screenNumberToDisplay}${convertNumberToLetter(questionNumber)}.`,
-    };
-  });
+  const displayQuestions = formatSurveyScreenQuestions(
+    activeScreen.length && activeScreen[0].questionType === 'Instruction'
+      ? activeScreen.slice(1)
+      : activeScreen,
+    screenNumberToDisplay,
+  );
+
+  const isReviewScreen = !screenNumber;
+  let screenHeader = '';
+  if (activeScreen.length && activeScreen[0].questionText) {
+    screenHeader = activeScreen[0].questionText;
+  }
 
   return (
     <SurveyFormContext.Provider
@@ -133,9 +118,10 @@ export const SurveyContext = ({ children }) => {
         isLast,
         numberOfScreens,
         screenNumber,
+        isReviewScreen,
         displayQuestions,
         surveyScreenComponents,
-        screenHeader: activeScreen.length ? activeScreen[0].questionText : '',
+        screenHeader,
       }}
     >
       <SurveyFormDispatchContext.Provider value={dispatch}>
