+++ packages/datatrak-web/src/layout/ScrollableBody.tsx
@@ -1,0 +1,14 @@
+/*
+ * Tupaia
+ *  Copyright (c) 2017 - 2023 Beyond Essential Systems Pty Ltd
+ */
+import styled from 'styled-components';
+
+export const ScrollableBody = styled.div`
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+  flex: 1;
+  overflow-y: auto;
+  padding: 1rem 2.5rem;
+`;
