+++ packages/datatrak-web/src/layout/index.ts
@@ -6,4 +6,5 @@ export { Header } from './Header';
 export { BackgroundPageLayout } from './BackgroundPageLayout';
 export { MainPageLayout } from './MainPageLayout';
 export { CentredLayout } from './CentredLayout';
-export { ScrollableLayout } from './ScrollableLayout';
+export { ScrollableBody } from './ScrollableBody';
+export { SurveyLayout } from './SurveyLayout';
