+++ packages/datatrak-web/src/theme/theme.ts
@@ -49,6 +49,7 @@ export const theme = createMuiTheme({
       },
       h2: {
         fontSize: '1.125rem', // survey page titles
+        fontWeight: 600,
       },
     },
   },
