+++ packages/datatrak-web/src/views/SurveyPage.tsx
@@ -34,6 +34,7 @@ export const SurveyPage = () => {
     <SurveyContext>
       <PageWrapper>
         <SurveyToolbar />
+
         {/* Use a key to render a different survey screen component for every screen number. This is so
       that the screen can be easily initialised with the form data. See https://react.dev/learn/you-might-not-need-an-effect#resetting-all-state-when-a-prop-changes */}
         <Outlet key={screenNumber} />
